move js files from frontend to javascript
