+++ app/javascript/packs/password_check.js
