+++ app/javascript/packs/utilities/password_confirmation.js
