+++ app/javascript/packs/utilities/sorted_table.js
